requirements.txt
fastapi
uvicorn
google-generativeai
jinja2